% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get a token from arcgis.com}
\usage{
get_token(token_url = "https://www.arcgis.com/sharing/rest/generateToken")
}
\arguments{
\item{token_url}{Optional. The url the token will be retrieved from.}
}
\description{
\code{get_token} retrieves a token from arcgis.com with the supplied username and password and stores them in \code{options}.
}
